requirements.txt
asn1crypto==0.24.0
astroid==2.0.4
beautifulsoup4==4.6.3
birdy==0.3.2
certifi==2018.8.24
cffi==1.11.5
chardet==3.0.4
Click==7.0
cryptography==2.3.1
cssselect==1.0.3
cycler==0.10.0
feedfinder2==0.0.4
feedparser==5.2.1
Flask==0.12.2
Flask-Login==0.4.0
Flask-Session==0.3.1
Flask-SocketIO==3.0.2
idna==2.7
isort==4.3.4
itsdangerous==0.24
jieba3k==0.35.1
Jinja2==2.7.3
kiwisolver==1.0.1
lazy-object-proxy==1.3.1
lxml==4.2.5
MarkupSafe==0.23
matplotlib==3.0.0
mccabe==0.6.1
newspaper3k==0.2.8
nltk==3.4
numpy==1.15.2
oauthlib==2.1.0
pandas==0.23.4
Pillow==5.3.0
pkg-resources==0.0.0
pony==0.7.6
pycparser==2.19
pylint==2.1.1
pymongo==3.7.2
PyMySQL==0.9.2
pyparsing==2.2.2
python-dateutil==2.7.3
python-engineio==2.3.2
python-socketio==2.0.0
pytz==2018.5
PyYAML==3.13
requests==2.19.1
requests-file==1.4.3
requests-oauthlib==1.0.0
selenium==3.14.1
singledispatch==3.4.0.3
six==1.11.0
tinysegmenter==0.3
tldextract==2.2.0
typed-ast==1.1.0
Unidecode==1.0.22
urllib3==1.23
Werkzeug==0.12.2
wrapt==1.10.11
